[Feat] tabToday (directory)
'작성하기'
'작성하기' 버튼을 누르면 답변 작성 뷰로 넘어가도록 구현했습니다.
또 리오의 피드백을 반영하여 리팩토링을 진행하였습니다.
